requirements.txt
PyQt6
google-genai
protobuf
google-api-core
google-auth
markdown
Pygments